+++ src/components/pages/apps/network/network-svg.tsx
@@ -5,7 +5,10 @@ import { SvgBase } from '@/components/plot/svg-base'
 import { SvgMargin } from '@/components/plot/svg-margin'
 
 import { useEdbSettings } from '@/components/edb/edb-settings'
+import { SvgCircle } from '@/components/plot/svg-circle'
 import { SvgG } from '@/components/plot/svg-g'
+import { SvgLine } from '@/components/plot/svg-line'
+import { SvgText } from '@/components/plot/svg-text'
 import { COLOR_BLACK } from '@/lib/color/color'
 import { useTooltip } from '@/providers/tooltip-provider'
 import { useZoom } from '@/providers/zoom-provider'
@@ -70,56 +73,154 @@ export function NetworkSvg() {
         .flat()
     )
 
-    console.log(groupMap)
+    console.log(groupMap, settings.plot.margin)
 
     const svg = (
-      <SvgMargin margin={settings.plot.margin}>
-        {network.edges.map((edge, idx) => {
-          const sourcePos = coordinates[edge.source] || { x: 0, y: 0 }
-          const targetPos = coordinates[edge.target] || { x: 0, y: 0 }
-          return (
-            <line
-              key={idx}
-              x1={sourcePos.x}
-              y1={sourcePos.y}
-              x2={targetPos.x}
-              y2={targetPos.y}
-              stroke="#cbd5e1"
-              strokeWidth={edge.score * settings.plot.edges.scale}
-            />
-          )
-        })}
+      <>
+        <SvgMargin margin={settings.plot.margin}>
+          <rect
+            x={0}
+            y={0}
+            width={size.w}
+            height={size.h}
+            fill="none"
+            stroke="black"
+          />
+          {settings.plot.edges.line.show &&
+            network.edges.map((edge, idx) => {
+              const sourcePos = coordinates[edge.source] || { x: 0, y: 0 }
+              const targetPos = coordinates[edge.target] || { x: 0, y: 0 }
+              return (
+                <SvgLine
+                  key={idx}
+                  x1={sourcePos.x}
+                  y1={sourcePos.y}
+                  x2={targetPos.x}
+                  y2={targetPos.y}
+                  s={settings.plot.edges.line}
 
-        {network.nodes.map((node) => {
-          const pos = coordinates[node.id] || { x: 0, y: 0 }
+                  strokeWidth={edge.score * settings.plot.edges.scale}
+                />
+              )
+            })}
 
-          let color = COLOR_BLACK
+          {network.nodes.map((node) => {
+            const pos = coordinates[node.id] || { x: 0, y: 0 }
 
-          switch (settings.plot.nodes.color.mode) {
-            case 'group':
-              console.log(node.group, groupMap[node.group.toLowerCase()])
-              color = groupMap[node.group.toLowerCase()]?.color ?? COLOR_BLACK
-              break
+            let fillColor = COLOR_BLACK
 
-            default:
-              color = COLOR_BLACK
-              break
-          }
+            switch (settings.plot.nodes.color.mode) {
+              case 'group':
+                fillColor =
+                  groupMap[node.group.toLowerCase()]?.color ?? COLOR_BLACK
+                break
 
-          return (
-            <SvgG key={node.id} pos={pos}>
-              <circle r={node.size * settings.plot.nodes.scale} fill={color} />
-              <text textAnchor="middle" dy=".3em" fill="#fff" fontSize={11}>
-                {node.label}
-              </text>
+              default:
+                fillColor = COLOR_BLACK
+                break
+            }
+
+            if (node.label === 'BLOOD MODULE-1.4 UNDETERMINED') {
+              console.log('found node:', node, pos)
+            }
+
+            let textAnchor: 'start' | 'middle' | 'end' = 'middle'
+
+            switch (settings.plot.nodes.labels.position) {
+              case 'left':
+                textAnchor = 'start'
+                break
+              case 'center':
+                textAnchor = 'middle'
+                break
+              case 'right':
+                textAnchor = 'end'
+                break
+              case 'below':
+              case 'above':
+                textAnchor = 'middle'
+                break
+              default:
+                textAnchor = 'middle'
+                break
+            }
+
+            return (
+              <SvgG key={node.id} pos={pos}>
+                <circle
+                  r={node.size * settings.plot.nodes.scale}
+                  fill={fillColor}
+                  fillOpacity={settings.plot.nodes.color.opacity}
+                />
+                <SvgText
+                  textAnchor={textAnchor}
+                  //dy=".3em"
+                  font={settings.plot.nodes.labels.text}
+                  fill={
+                    settings.plot.nodes.labels.color.on
+                      ? fillColor
+                      : settings.plot.nodes.labels.color.default
+                  }
+                >
+                  {node.label}
+                </SvgText>
+              </SvgG>
+            )
+          })}
+        </SvgMargin>
+        <SvgG
+          pos={{
+            x: settings.plot.margin.left + size.w + 20,
+            y: settings.plot.margin.top,
+          }}
+        >
+          <SvgG id="group-legend">
+            <SvgText
+              textAnchor="start"
+              font={settings.plot.nodes.labels.text}
+              fill={COLOR_BLACK}
+              fontWeight="bold"
+            >
+              Groups
+            </SvgText>
+            <SvgG
+              pos={{ x: settings.plot.legend.dot.radius, y: 10 }}
+              id="groups"
+            >
+              {groups.map((group, gi) => (
+                <SvgG
+                  key={group.id}
+                  pos={{
+                    x: 0,
+                    y: 10 + gi * (settings.plot.legend.dot.radius * 2 + 5),
+                  }}
+                >
+                  <SvgCircle
+                    r={settings.plot.legend.dot.radius}
+                    fill={group.color}
+                    stroke={COLOR_BLACK}
+                  />
+                  <SvgG pos={{ x: settings.plot.legend.dot.radius + 5, y: 0 }}>
+                    <SvgText
+                      textAnchor="start"
+                      font={settings.plot.nodes.labels.text}
+                      fill={COLOR_BLACK}
+                    >
+                      {group.name}
+                    </SvgText>
+                  </SvgG>
+                </SvgG>
+              ))}
             </SvgG>
-          )
-        })}
-      </SvgMargin>
+          </SvgG>
+        </SvgG>
+      </>
     )
 
     return { svg, width, height }
   }, [settings, size, network, coordinates])
+
+  console.log('bee  ', width, height)
 
   return (
     <SvgBase width={width} height={height} scale={zoom}>