+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- More props for network rendering.
 - Added support for groups with color to network.